moving enum to different file
