+++ src/views/htmlcontent/src/js/services/data.service.ts
@@ -8,6 +8,7 @@ import { ISlickRange } from 'angular2-slickgrid';
 import { QueryEvent, ResultSetSubset, ISelectionData } from './../../../../../models/interfaces';
 import * as Constants from './../constants';
 import { createProxy, IMessageProtocol, IServerProxy } from '../../../../../protocol';
+import { TelemetryActions, TelemetryViews } from '../../../../../telemetry/telemetryInterfaces';
 
 declare function acquireVsCodeApi(): { postMessage: (message: string) => void; };
 
@@ -85,8 +86,13 @@ export class DataService implements OnDestroy {
 		this._proxy.saveResults(batchIndex, resultSetNumber, format, selection);
 	}
 
-	sendSimpleTelemetryEvent(view: string, action: string, properties?: { [key: string]: string }, measurement?: {[key: string]: number} ): void {
-		this._proxy.sendSimpleTelemetryEvent(view, action, properties, measurement);
+	sendActionEvent(
+		telemetryView: TelemetryViews,
+		telemetryAction: TelemetryActions,
+		additionalProps: { [key: string]: string },
+		additionalMeasurements: { [key: string]: number }
+		): void {
+		this._proxy.sendActionEvent(telemetryView, telemetryAction, additionalProps, additionalMeasurements);
 	}
 
 	/**
